ecflow_server: correct default status log optimization
These changes correct the optimisation/inhibition of unnecessarily logging the node state change, applicable to the descendants of a node with default status COMPLETE.

Without these changes, the controlling flag was update+used and, because it was passed as reference, later incorrectly applied to siblings nodes.

To avoid similar issues in the future, all Requeue_args data members not currently updated in the scope of handling requeue are made constant.

Re ECFLOW-1914