Replace checkpoint file atomically
These changes modify the strategy used to store a checkpoint file.

Previously the storage operation consisted of (a) moving the current
checkpoint file to a backup, and then (b) storing the contents of the
checkpoint file. If the server was terminated during (b), the result
would be a corrupt checkpoint file.
The new strategy consists of (a) storing the contents of the checkpoint
file in a temporary file; (b) moving the current checkpoint to backup;
and, finally, (c) promoting the temporary checkpoint file to the current
by simply renaming the file. In this case, even if the storage operation
is interrupted, the current checkpoint file will never lack integrity.

Re ECFLOW-1925