Immediately free task held by Aviso attributes
This change implies that, once a notification by Aviso is detected, the AbstractServer::job_gen_count is incremented.

It is important to notice that this counter determines if the NodeTreeTraversal actually traverses the node tree to check if any attribute has free'd the associated node and job generation+creation can happen.

Re ECFLOW-1931