Correct parsing of Label attribute
The changes propose a better (but still not flawless) approach to
parsing the Label attribute.

Re ECFLOW-1890

Explanation of the Problem:

A Label attribute is stores in the .def file in the following format:

> label <name> "<value>" # "<new-value>"

with the double quotes and hash symbol used as separators.
This approach works fine while <value> or <new-value> do not contain
any double quotes or hash symbols.

When the values contain double quotes or hash symbols, the parsing
becomes ambiguous. In the following example

> label <name> "value X " # " and Y " # " value Z"

Should the parser consider
  <value>     = "value X " # " and Y "
  <new-value> = " value Z"
or
  <value> = "value X "
  <new-value> " and Y " # " value Z"
?

Both might be correct depending on the user's intents!

This becomes even more convoluted when the value contains double quotes,
as in this example:

label <name> " "value X" " # " # " # " "value Y" "